+++ .gitignore
@@ -1,3 +1,8 @@
+# tests
+tests/services/jupiter_crawler/html_contents
+tests/services/jupiter_crawler/new_results.json
+tests/services/jupiter_crawler/results.json
+
 # build
 build/
