+++ server/utils/enums/class_type.py
@@ -1,19 +1,26 @@
 from enum import Enum
-from typing import Self
 
 
 class ClassType(Enum):
     PRACTIC = "practic"
     THEORIC = "theoric"
+    VINCULATED_THEORIC = "vinculated_theoric"
+    VINCULATED_PRACTIC = "vinculated_practic"
 
     @classmethod
     def from_str(cls, value: str) -> "ClassType":
         pratic_values = ["Prática"]
         teoric_values = ["Teórica"]
+        vinculated_theoric_values = ["Teórica Vinculada"]
+        vinculated_practic_values = ["Prática Vinculada"]
         if value in pratic_values:
             return cls.PRACTIC
         if value in teoric_values:
             return cls.THEORIC
+        if value in vinculated_theoric_values:
+            return cls.VINCULATED_THEORIC
+        if value in vinculated_practic_values:
+            return cls.VINCULATED_PRACTIC
         raise NoSuchClassType(f"Class type {value} is not valid.")
