+++ tests/services/jupiter_crawler/crawler_test_cases.txt
@@ -1,0 +1,77 @@
+0323201
+4323101
+MAP3121
+MAP3122
+PCS3100
+PCS3111
+PCS3115
+PCS3225
+PCS3519
+PCS3528
+PCS3539
+PCS3544
+PCS3548
+PCS3563
+PCS3616
+PCS3623
+PCS3818
+PEA3100
+PEA3301
+PEA3306
+PEA3311
+PEA3388
+PEA3390
+PEA3391
+PEA3395
+PEA3400
+PEA3402
+PEA3410
+PEA3411
+PEA3413
+PEA3420
+PEA3450
+PEA3489
+PEA3496
+PEA3503
+PEF3208
+PMT3100
+PMT3131
+PRO3811
+PRO3821
+PSI3211
+PSI3213
+PSI3321
+PSI3322
+PSI3421
+PSI3431
+PSI3441
+PSI3451
+PSI3471
+PSI3481
+PSI3531
+PSI3541
+PSI3552
+PSI3560
+PSI3561
+PSI3581
+PTC3007
+PTC3307
+PTC3405
+PTC3413
+PTC3415
+PTC3419
+PTC3422
+PTC3424
+PTC3432
+PTC3440
+PTC3441
+PTC3444
+PTC3450
+PTC3520
+PTC3527
+PTC3545
+PTC3547
+PTC3548
+PTC3566
+PTC3575
+PTC3578
