+++ tests/services/jupiter_crawler/test_crawler.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+@pytest.mark.asyncio
+async def test_crawler() -> None:
+    pass
